* 'master' of zoidberg.inventorum.net:inventorum/ebay: (22 commits)
  - version 0.1.8
  Add workaround for deletions
  Document CHANGES
  Use `inv_product_id` to lookup updated core products
  Bumped to dev version
  Bumped version
  Accept null for variations inv_id
  String format, check
  Fix environment settings for rc and staging
  Fix ebay ru for rc
  Fix failing test because of auto-merge (value doens't matter here :-)
  Use inv_id whereever possible + fix tests
  Add inv_product_id to EbayItemModel + migrations
  Fix failing tests due to new image model structure
  Document CHANGES
  Avoid double publishing
  Add failing test
  Cleanup
  Changed ebay runame for staging
  - commands for db creation
  ...

+++ src/inventorum.ebay/inventorum/ebay/lib/db/models.py
@@ -43,7 +43,7 @@ class MappedInventorumModel(BaseModel):
     class Meta:
         abstract = True
 
-    inv_id = models.IntegerField(unique=True, verbose_name="Universal inventorum id")
+    inv_id = models.BigIntegerField(unique=True, verbose_name="Universal inventorum id")
 
     objects = PassThroughManager.for_queryset_class(MappedInventorumModelQuerySet)()
